// src/app/interfaces/leader-board.ts
export interface LeaderBoard {
  points001Best: Array<{ value: number; name: string }>;
  points002Best: Array<{ value: number; name: string }>;
  points003Best: Array<{ value: number; name: string }>;
  points004Best: Array<{ value: number; name: string }>;
  points005Best: Array<{ value: number; name: string }>;
  points006Best: Array<{ value: number; name: string }>;
  points007Best: Array<{ value: number; name: string }>;
  points008Best: Array<{ value: number; name: string }>;
  points009Best: Array<{ value: number; name: string }>;
  points010Best: Array<{ value: number; name: string }>;
  points011Best: Array<{ value: number; name: string }>;
  points012Best: Array<{ value: number; name: string }>;
  points013Best: Array<{ value: number; name: string }>;
  points014Best: Array<{ value: number; name: string }>;
  points015Best: Array<{ value: number; name: string }>;
  points016Best: Array<{ value: number; name: string }>;
  points017Best: Array<{ value: number; name: string }>;
  points018Best: Array<{ value: number; name: string }>;
  points019Best: Array<{ value: number; name: string }>;
  points020Best: Array<{ value: number; name: string }>;
}